Extract institution classification logic into a custom hook
Refactored DataDetails component to remove redundant logic by moving institution classification-related functionality into the new useInstitutionClassificationDataHandlers hook. This improves code reusability, readability, and maintainability.
